Simplify module structure to libdependency_graph library
So that it can be tested easier from command line.
