+++ thoth/adviser/python/bin/__init__.py
@@ -18,6 +18,6 @@
 """Interface to libdependencygraph.so."""
 
 from .dependency_graph import DependencyGraph
-from .exceptions import DependencyGraphException
-from .exceptions import PrematureStreamEndError
-from .exceptions import NoDependenciesError
+from .dependency_graph import DependencyGraphException
+from .dependency_graph import PrematureStreamEndError
+from .dependency_graph import NoDependenciesError
